+++ packages/sanity/src/form/studio/StudioFormBuilder.test.tsx
@@ -5,7 +5,7 @@ import {defineType, Path} from '@sanity/types'
 import {render} from '@testing-library/react'
 import React, {useMemo} from 'react'
 import {createMockSanityClient} from '../../../test/mocks/mockSanityClient'
-import {createTestProvider} from '../../../test/form'
+import {createTestProvider} from '../../../test/testUtils/TestProvider'
 import {createPatchChannel} from '../patch'
 import {useFormState} from '../store/useFormState'
 import {EMPTY_ARRAY} from '../utils/empty'
